Extract use cases into separate module and abstract them with interfaces
