+++ OTAnalytics/application/analysis/traffic_counting_specification.py
@@ -1,0 +1,28 @@
+from abc import ABC
+from dataclasses import dataclass
+from typing import Iterable
+
+
+@dataclass(frozen=True)
+class ExportFormat:
+    name: str
+    file_extension: str
+
+
+@dataclass(frozen=True)
+class CountingSpecificationDto:
+    """
+    Data transfer object to represent the counting.
+    """
+
+    interval_in_minutes: int
+    format: str
+    output_file: str
+
+
+class ExportCounts(ABC):
+    def export(self, specification: CountingSpecificationDto) -> None:
+        raise NotImplementedError
+
+    def get_supported_formats(self) -> Iterable[ExportFormat]:
+        raise NotImplementedError
